Devel

+++ arduino/libraries/ambu_common/SensorHaf50Slpm.cpp
@@ -12,16 +12,15 @@ SensorHaf50Slpm::SensorHaf50Slpm () : GenericSensor(HAF_50SLPM_ADDR) { }
 void SensorHaf50Slpm::update(uint32_t ctime) {
    uint8_t data[2];
    uint16_t x;
-   double raw;
-
+   uint16_t raw;
    // Read value
    Wire.requestFrom(addr_, byte(2));
    for (x=0; x < 2; x++) data[x] = Wire.read();
 
-   raw = (double)((data[0] << 8) | data[1]);
+   raw = (data[0] << 8) | data[1];
 
    // Scaled value
-   scaled_ = 50.0 * (((raw / 16384.0) - 0.1) / 0.8);
+   scaled_ = 50.0 * (((double(raw) / 16384.0) - 0.1) / 0.8);
 }
